Split paths per-OS bases into cfg-selected imp files
config/cache/log getters now own the shared policy (override, ensure,
app-dir join, log-to-file); each OS contributes only its base dirs in
its own file, selected by one cfg_attr module. Replaces the 11-attr
cfg cluster the per-getter triplication produced.
